Return JSON with the organisation slugs
Exercises the creation of the Organisation model along with its
controller. We keep the scope of this commit small, and will add the
ContentItems model in future commits.

We are using a very basic mechanism to serialise to JSON the
organisations .
This is because most of this code will go away once we expose
an API to other apps. We will need to rely on ApiSerializers[2] or
using libraries like Grape[1]. No need to complicate things further.

[1]: https://github.com/ruby-grape/grape
[2]: https://github.com/rails-api/active_model_serializers