Add rake task to import organisation
The task receives the slug of the organisation and import the Content
Items. It delegates the importing process to Importers::Organisations,
as the rake task is mainly a way to import the data, but there can be
other ways, for example, using `rails console` directly.

**Developer Note:**

 `Reenable` rake task before each tests run.

Invoke does not run the task twice, for example if you invoke the
same task in a previous test it won't run again.

So reenable call is required to allow tasks to run if the task
is invoked again.

See this url for confirmation [1]

[1] http://apidock.com/ruby/v1_9_3_392/Rake/Task/reenable