Add number of content items to Organisation JSON
To count the number of content items per organisation we need to create
the model (ContentItem) along with setting up the associations.

To facilitate testing we are including the library Shoulda[1]

We are using a very basic mechanism to serialise to JSON the
organisations and the requested number of Content Items.
This is because most of this code will go away once we expose
an API to other apps. We will need to rely on ApiSerializers[3] or
using libraries like Grape[2]

It is not ideal either the way we are mocking the HTTP requests, as we
are not webmoking the get request, but the library. Again, this is
something that will go away in a few weeks, so no need to work any
further.

[1]: https://github.com/thoughtbot/shoulda-matchers
[2]: https://github.com/ruby-grape/grape
[3]: https://github.com/rails-api/active_model_serializers