Import an organisation and ContentItems via SearchAPI
Creates the Organisation with the provided slug. It queries the
searchAPI with and filters out all ContentItems per organisation.

For each ContentItem it stores its `content_id`. This is, so far, the
only attribute that we need to store in the model for current story. In
future stories we will evolve it to include other attributes.

We temporarily rely on [HTTParty][1] to query SearchAPI; this is a
temporary solution as we will use GDS  Adapters as a way to communicate
with other applications. The task to upgrade to use GDS adapters is
tracked in a [Trello Card][2]

It is not ideal either the way we are mocking the HTTP requests, as we
are not webmoking the get request, but the library. Again, this is
something that will go away in a few weeks, so no need to work any further.

[1]: https://github.com/jnunemaker/httparty
[2]: https://trello.com/c/mn8JPHYR/78-use-gds-adapter-to-retrieve-content-items-for-an-organisation