Import an organisation and ContentItems via SearchAPI
Creates the Organisation with the provided slug. It queries the
searchAPI with and filters all ContentItems per organisation.

For each ContentItem it stores its `content_id`. This is, so far, the
only attribute that we need to store in the model for current story. In
future stories we will evolve it to include other attributes.

We temporarily rely on [HTTParty][1] to query SearchAPI; this is a
temporary solution as we will use GDS  Adapters as a way to communicate
with other applications. The task to upgrade to use GDS adapters is
tracked in a [Trello Card][2]

It is not ideal either the way we are mocking the HTTP requests, as we
are not webmoking the get request, but the library. Again, this is
something that will go away in a few weeks, so no need to work any further.

Raises a "No result for slug" exception so the caller understands when
and why the organisation has not been loaded: Every Organisation should
have, at least, a ContentItem. If during the migration process, we found
 an organisation with no Content Items we raise an exception as there is
a content problem that need to be fixed for that organisation.

[1]: https://github.com/jnunemaker/httparty
[2]: https://trello.com/c/mn8JPHYR/78-use-gds-adapter-to-retrieve-content-items-for-an-organisation