Raise exception when results is empty
Raises a "No result for slug" exception so the caller understands when
and why the organisation has not been loaded.

The reason for the error is the following:

Every Organisation should have, at least, a ContentItem. If during the
migration process, we found an organisation with no Content Items we
raise an exception as there is a content problem that need to be fixed
for that organisation.