Renames app to content-performance-manager
When we set up the Repo in Github is wasn’t clear to us what the name
of the application were going to be, so we decided to go with a
temporary one: com-prototype.

But, we have agreed to name it `content-performance-manager`, so it is
time to update the name across the Rails app.

This commit updates all references to the old name (com-prototype) 
across the codebase:

- Layout
- Application config
- Database names
- Session info