Add a rake task to configure the Heroku remotes
**Note:**
Deploying to Heroku is the temporary option we are going to use while
a more robust is configured for the application.

To deploy to Heroku we need to setup the remotes, so this task includes
set up the dev environment to allow deployments to production.

Enables this rake task:

```bash
$ bundle exec rake heroku:prepare
```