Include the migration when deploying to Heroku
It feels more convenient for a developer not to have to run the
database migration each time he needs to do a deployment.

For this reason, I am including the migration task within the deployment
task.

As a side note, I had to add the `Bundler.with_clean_env` to run the
migrations because Heroku was picking up the wrong version of ruby.
See suggested solution in StackOverflow[1] and explanation in:

[1]: http://stackoverflow.com/questions/12995271/rails-server-error-ruby-version-is-1-8-7-but-your-gemfile-specified-1-9-3/20668794#20668794