Render content item’s url
For each department’s content item, content designers want to visit the url of
a content item by clicking a link.

This new method (i.e url) has been introduced  concatenates https://gov.uk and
the content item’s path.

This is then used to populate a Content URL column, Ideally this url will be
wrapped in the content title. but for the purpose of this story it has been
wrapped with the url