Remove explicit logging configuration.
The rake task outputs content to STDOUT, but if we use it from the
console we would like to use the defaults configured in
`config/environments/development.rb`.
Likewise, if we are in  production, we would like to use the
configuration specified in `config/environments/production.rb`.

To summarise it, we want to get rid of any configuration code related
to the the current environment.

```ruby
# We avoid writing this
if Rails.env == 'test'
  # configure loggin
end
```

We can configure logging per environment in the files:

- config/environments/test.rb
- config/environments/development.rb
- config/environments/production.rb

## Developer note:

If the logging information is valuable for the script user we will
store it in a more reliable place (Database, filesytem...)
We have chosen a quick way to get feedback from the rake task until we
have a better picture about all the issues we have when parsing the data