Improves readability for `public_updated_at`
From the user perspective it is easier to read `2 months ago` than the
exact moment in time when something has been updated.
It feels more beneficial for the user to get a quick idea of how old a 
piece of content is by reading a text description of the current gap.