Render ascending and descending links
When a user clicks on the last update table column the
ordering of the column changes to ascending or descending.

Here helper method is used to updated the ordering links.
When no ordering is provided, the link defaults to descending,
when descending is supplied the link produced is for ascending,
vice versa.

These have been included to help with sorting the column by the
last updated time.