Don't create an organisation if it already exists
The import script was creating an organisation every time it ran. There were a
few reason for this.

1) the autogenerated id was the only unique field
2) no checks were being done against the database to see if the organisation
already exists.

A constraint has been added to the database to make the `slug` column unique.

In order to stop the new uniqueness constraint in the database from throwing an
exception, the import task checks whether an organisation already exists before
trying to create it.