Update content item if it already exists
As with organisations, the import task did not check whether a content item
exists before adding it to the database. This wasn't an issue before, when
every time the import ran, a new organisation was created.

Now we would run the risk of an organisation having many duplicated content
items.

Added a uniqueness constraint to the content_id field. This field is not likely
to change as it is a guid that is returned by the search api for each content
item (or at the very least, should be returned for each content item) and should
already be unique.

Added a check to see if a content item already exists for an organisation, and
if so, just update all of it’s attributes.

One of the pagination tests also needed to be updated. As we no longer allow
duplicate content items to be created, all of the mocked responses have
to be unique as well.