Add a logger parameter to Importers::Organisation
Making the logger configurable via parameter allow us to be specific
about what output type we want in the rake task without being
environment dependent.

## Developer note:

If the logging information is valuable for the user of the script we
will store it in a more reliable place (Database, filesytem...) We have
chosen a quick way to get feedback from the rake task until we have a
better picture about all the issues we have when parsing the data.