Refactor `Importers::Organisation` rake task test
Use `instance_double`[1] to verifies that any methods being stubbed
would be present on an instance of that class. In  addition, when it
receives messages, it verifies that the provided arguments are supported
by the method signature, both in terms of arity and allowed or required
keyword  arguments, if any.

Extract a description context for the constructor parameters so the
description  of the RSpec object is more clear.

```text
Randomized with seed 22688

Import organisation rake task
  import:organisation
    runs the process to import the organisations
    Importer parameters
      receives the slug
      receives the logger
      raises an error if a slug parameter is not present

Finished in 0.12665 seconds (files took 5.65 seconds to load)
4 examples, 0 failures
```

[1]: https://relishapp.com/rspec/rspec-mocks/v/3-2/docs/verifying-doubles/using-an-instance-double