Fix SASS workaround to not use unnecessary `!optional`
When Elements was initially implemented (in #5)
it was throwing an error which was fixed via a hack,
i.e. using `!optional` when `@extend`ing something.

This fixes it, as it turned out that `= require_tree .`
includes CSS files in an unspecified order, meaning
dependencies are not always met.

This also corrects `base.css` to be a partial.