Display the total number of content items.
The original acceptance criteria asks for the total number of pages to be
displayed, however the Kaminari gives the count of items per page out of the
box without any extra configuration.

e.g. "Displaying content items 1 - 2 of 3 in total"

For a first pass, I think this will do.

If the total number of pages really is needed by the users, we'll have to
either find a way to use Kaminari's `total_pages` function [1], or write
one of our own. See this link [2] for examples.

[1](https://github.com/amatsuda/kaminari/blob/8763cad3d71e53477c822b340607f1507118ecf4/kaminari-core/lib/kaminari/models/page_scope_methods.rb)
[2] https://github.com/amatsuda/kaminari/blob/master/kaminari-core/lib/kaminari/helpers/