Import the organisation title
The organisation title doesn’t always match the slug of the
organisation. If we want to make that we are displaying the correct
organisation title to the user, we should retrieve it from the search
api.

The search api returns a list of organisations that each content item
belongs to. As we are filtering the search api by organisation, this list
only contains one item, so it is safe to only ask for the first
organisation and save it’s title.

At the moment the organisation title is being saved every time a content
item is saved. This will change in the next commit, where organisation
information will only be stored once per import run.