Only update the organisation title once per import
As the organisation title is being retrieved from the content item data
it’s possible that it could be updated thousands of times per organisation.

Add a check to make sure the organisation title is only update once.

Some substantial changes have been made to the import `run` function to get this
to work:

1) The code to update the organisation was moved to a function to make it easier
to test. It's easier to check how many times a function has been called, rather
than a database operation.

2) The `organisation` variable was changed to an instance variable to make it
accessible outside of the `run` function.

3) Most importantly, the location of code that calls the update to the
organisation has moved. As the organisation title is retrieved from the content
item data, the code was originally in the `each` loop that loops over the content
items. Unfortunately, for reasons I haven't been able to determine, in this
location, the update to `@organisation` returns nil.

We know that the database is object is being updated, as the test added in the
previous commit confirms that. However the `@organisation` instance variable is
not, so `@organisation.title.blank?` always returns `true`.

Moving the code that calls the organisation to above the `each` loop is not
elegant, but it works.