Render the organisation title
Render the organisation title above the list of content items. The
organisation title replaces the name that was generated from the 
organisation slug.

Removed the `name` function in the model because it shouldn’t be 
possible to retrieve an organisation from the search api that doesn’t
have a title.