Add document type for a content item
For every content item, content designers want to see the type of document a
content item is belonging to an organisation.

Here a new column/attribute (i.e `document_type`) has been added to the
content_item(s).

The `document_type` attribute’s value is obtained from the content item
store [1].

The content item end point provides access to the document_type JSON property

For obtaining the document type a piece of content_item, this endpoint [1] is queried
with the base_path supplied.

[1] /api/content/:content_item_base_path