Extract Client to load Content Item details
Content Items details are fetched from the Content Store. so I have
extracted a Client that knows how to connect to the Content Store, and
parse the response.

This commit is part of a major refactor (PR 38). The problem it is
trying to fix is: `::Importers::Organisation` has many responsibilities,
and has grown without much control. It is  difficult to test and
maintain,because it knows about the internals of Content Store and
Search API (both API and data transformation).

See PR 38 for additional context.