Extract Client to fetch Content Item by Organisations
Search API returns the Content Items belonging to an organisation. I
have created a class to connect to the Search API and parse the
response so other classes are not coupled to the remote service.

This commit is part of a major refactor (PR 38). The problem it is
trying to fix is: `::Importers::Organisation` has many responsibilities,
and has grown without much control. It is  difficult to test and
maintain,because it knows about the internals of Content Store and
Search API (both API and data transformation).

See PR 38 for additional context.