Use instance methods in Clients::SearchAPI
There is state that is being shared among objects as parameters in
the method parameters list. To prevent this situation I am using instance
methods as a preparation to store state in instance variables.

This commit is part of a major refactor (PR 38). The problem it is
trying to fix is: `::Importers::Organisation` has many responsibilities,
and has grown without much control. It is  difficult to test and
maintain,because it knows about the internals of Content Store and
Search API (both API and data transformation).

See PR 38 for additional context.