Remove `start` parameter in Clients::SearchAPI:fetch
I have isolated the client to connect to the Search API, so we don't
really need to expose the start parameter in the client API.

This commit is part of a major refactor (PR 38). The problem it is
trying to fix is: `::Importers::Organisation` has many responsibilities,
and has grown without much control. It is  difficult to test and
maintain,because it knows about the internals of Content Store and
Search API (both API and data transformation).

See PR 38 for additional context.