Add a link to the content item `detail` page
Update the content_item view to add the link to content items detail page in the
table of content items.

Add an empty header to the content items table. This is because the
link name acts as a column heading.

Add an id to the content item in the content items factory. This allows the
content item id to be accessed in the view tests.