Add rake task to import all organisations
Importer::Organisations is not doing a very good job in retrieving the
information about an organisation, because it is reusing an existing get
request

I am using the following request to populate the list of organisations:
https://www.gov.uk/api/search.json?filter_format=organisation

This commit is part of a major refactor (PR 38). The problem it is
trying to fix is: `::Importers::Organisation` has many responsibilities,
and has grown without much control. It is  difficult to test and
maintain,because it knows about the internals of Content Store and
Search API (both API and data transformation).

See PR 38 for additional context.