Use Collectors::ContentItems to fetch Content Items
Relies on SearchAPI client to query content items. This step will make
easier to reuse the client to perform additional queries (for example,
organisations)

This commit is part of a major refactor (PR 38). The problem it is
trying to fix is: `::Importers::Organisation` has many responsibilities,
and has grown without much control. It is  difficult to test and
maintain,because it knows about the internals of Content Store and
Search API (both API and data transformation).

See PR 38 for additional context.