Setup Google Analytics request client
In order to interact with the Google Analytics Reporting API, this defines
a wrapper service that sets up the a client, an instance of
AnalyticsReportingService with the google service account credentials
supplied.

At the moment the service account credentials is provided via a JSON file
that contains a private key, account’s email (i.e client’s email) and other
attributes.

Upon investigation [1] the private key, client email and scope appear to be
the most sort after, hence why the tests ensures that the api client’s
authorisation is setup as expected.

[1] https://github.com/google/google-auth-library-ruby/blob/82a4f8520cc8897bbcf0485d4ea678317619c84e/lib/googleauth/service_account.rb#L57