Render content item’s number of page views
For each organisation’s content item, content designers want to know the
number of page views of a content item.

In previous commits the number of page views attribute was defined and
added to the import.

This commit updates the content items view table to show the number of
page views (i.e `number_of_views`).