Rename collectors into services
I prefer to use services instead of collectors. It feels more consistent
with the current architecture.

This commit is part of a major refactor (PR 38). The problem it is
trying to fix is: `::Importers::Organisation` has many responsibilities,
and has grown without much control. It is  difficult to test and
maintain,because it knows about the internals of Content Store and
Search API (both API and data transformation).

See PR 38 for additional context.