Move clients inside services folder
Clients are currently used by services, and it makes sense to not expose
them in other parts of the application.

Developer Note: Clients will go away shortly as we will use GOVUK
adaptors.

This commit is part of a major refactor (PR 38). The problem it is
trying to fix is: `::Importers::Organisation` has many responsibilities,
and has grown without much control. It is  difficult to test and
maintain,because it knows about the internals of Content Store and
Search API (both API and data transformation).

See PR 38 for additional context.