Extract method in Importers::AllOrganisations
Also updates copy in RSpec tests

```
Importers::AllOrganisations
  #import_all
    when the organisation already exists
      updates the attributes
      does not create a new one
    when the organisation does not exist
      updates the attributes
      creates an organisation per attribute group
```

This commit is part of a major refactor (PR 38). The problem it is
trying to fix is: `::Importers::Organisation` has many responsibilities,
and has grown without much control. It is  difficult to test and
maintain,because it knows about the internals of Content Store and
Search API (both API and data transformation).

See PR 38 for additional context.