Extract method in Importers::ContentItemsByOrganisation
Also updates copy in RSpec tests

```
Importers::ContentItemsByOrganisation
  #run
    when the content item does not exist
      creates a content item per attribute group
      updates the attributes
    when the content item already exists
      updates the attributes
      does not create a new one
```

This commit is part of a major refactor (PR 38). The problem it is
trying to fix is: `::Importers::Organisation` has many responsibilities,
and has grown without much control. It is  difficult to test and
maintain,because it knows about the internals of Content Store and
Search API (both API and data transformation).

See PR 38 for additional context.