Dry-up Clients::SearchAPI
Also updates copy of RSpec tests

```
Clients::SearchAPI
  returns symbolized attributes
  builds the SearchAPI query
  when multile pages
    does not fail when last page has 0 results
    paginates through the response
```

This commit is part of a major refactor (PR 38). The problem it is
trying to fix is: `::Importers::Organisation` has many responsibilities,
and has grown without much control. It is  difficult to test and
maintain,because it knows about the internals of Content Store and
Search API (both API and data transformation).

See PR 38 for additional context.