Use keyword arguments in SearchAPI#find_each
It improves the readability of the code. It also includes some RSpec
copy changes.

The file .rubocop.yml has been updated because we are using keyword
arguments and rubocop requires to declare in the config file what is
the ruby version we are targeting.

This commit is part of a major refactor (PR 38). The problem it is
trying to fix is: `::Importers::Organisation` has many responsibilities,
and has grown without much control. It is  difficult to test and
maintain,because it knows about the internals of Content Store and
Search API (both API and data transformation).

See PR 38 for additional context.