Update documentation of rake tasks
Add documentation for two rake tasks. The first one to import all the
organisations, and the second one, to import all the content items for
an organisation.

* $ rake import:all_organisations
* $ rake import:content_items_by_organisation[{department-slug}]

This commit is part of a major refactor (PR 38). The problem it is
trying to fix is: `::Importers::Organisation` has many responsibilities,
and has grown without much control. It is  difficult to test and
maintain,because it knows about the internals of Content Store and
Search API (both API and data transformation).

See PR 38 for additional context.