Set up Errbit error reporting
Using an older version of the airbrake because after version 5, the
view tests fail with the following error:

```
Failure/Error: <%= sort_table_header "Title", "title" %>
     
ActionView::Template::Error:
  undefined method `sort_table_header' for #<#<Class:0x007fe658b743f0>:0x007fe6508420b8>
```

The Airbrake gem was re-written when it moved to version 5 and split
into two gems: `airbrake` and `airbrake-ruby`. Some investigation will
need to be done to determine why version 5 is not compatible with the
`sort_table_header` used to order the columns in the table on the
content items page.