Switch to unique page views
The query to Google Analytics passes `ga:pageviews` as the metric
expression.

From [pageviews vs unique_views](https://support.google.com/analytics/answer/1257084#pageviews_vs_unique_views):

A pageview is defined as a view of a page on your site that is being
tracked by the Analytics tracking code. If a user clicks reload after
reaching the page, this is counted as an additional pageview. If a user
navigates to a different page and then returns to the original page, a
second pageview is recorded as well.

A unique pageview, as seen in the Content Overview report, aggregates
pageviews that are generated by the same user during the same session.
A unique pageview represents the number of sessions during which that
page was viewed one or more times.

Change to using unique pageviews because it is a better metric to
discover the number of unique users visiting the page.