Move gems to test group
Jenkins does not install any gems that are in the `development` group, so the
master branch was failing to build because the linter and listen gems were
missing.

Moving the `listen` and `govuk-lint` gems from the `development` group to the
`development and test` group has fixed the problem because it allows the bundler
on Jenkins bundler to load `listen` and `govuk-lint` for the TEST environment.