ManytoMany: ContentItems<=>Organisations
The import fails when a content item appears to belong to more than one
organisations.

This commit seeks to re-defines the association between ContenItems and the
respective Organisations to a ManytoMany association.

This allows for content items to be associated with the relevant
organisations (i.e more than one organisation).

As a result of this re-definition of association some tests have failed,
requiring the updates to factory girl’s definition of organisation and
also similar updates to some tests and concrete code.

I have updated factory girl organisation_with_content_items. This has been
done to fix failing tests and provide insight on the nature of the relationship
between content item and organisation from a factory girl point.

This references [1].

[1] http://www.rubydoc.info/gems/factory_girl/file/GETTING_STARTED.md#Associations