Remove nested show for content item
Since we reworked the model to be many to many between organisations and
content items, the view for a single item within an organisation does
not need to be at a nested route. As in:

`/organisations/hm-revenue-customs/content_items/30339`

can just be

`/content_items/30339`

as that content item might also belong to another org at:

`/organisations/public-health-england/content_items/30339`

This change also displays the multiple organisations that an item belongs
to in the `content_items#show` view.