Add unfiltered content items view
We want users to be able to see an unfiltered list of Content Items.

The routes were already plumbed in from prior commit but the tests
were for specific aspects of the feature being implemented.

When we are in an unfiltered view, we will need to show for each
content item what organisations they belong to.

The index view for ContentItems has dual behaviour now which can be
encapsulated in a feature.

Make the controller do less work and abstract sorting params to the query
object. As the query object has moved the work to sort the items out of
the controller to the query object, this functionality should also be
tested in isolation from the controller.