Use a constant to set the default batch size
Made the batch size a constant rather than a attribute to make it easier to read
and change later.

Through trial and error, I’ve hit upon a batch size of 3000 for one
month' data. This size could be larger for 7 days page views, and smaller for
one year’s page views.