Fix display for no last updated at in list view
Where we have content items that have never been published, we are
attempting to apply a helper to nil values in the view, which is causing
server errors. To fix this we need to check in the view before attempting
to output anything for that field.

Includes a test that replicates the bug.