Fix display for no last updated at in show view
When we display details for a content item that has never been
published, we are attempting to apply a helper to a nil value in the
view, which is causing server errors. To fix this we need to check in the
view before attempting to output anything for that field.

Includes a test that replicates the bug.