Change structure page views in parsed response from GA
The parsed response from GA only returns one value for the page view, the
total number of page views in the last month.
We will need to change this in later commits to also return the total
number of page views for the last year.

As a preparatory step we are changing the page views object, to make it
easier to return multiple values for page views to a single content item.