Add delayed_job to page views import
We are seeing issues with timeouts when we try to fetch page views for
large sets of content items, so we are attempting to break the batches
into separate queues running as background jobs.

We moved the code that does the importing into the job. We also
refactor the tests as now the tests are exercised against the job's
code rather than for the importer itself. The importer now just calls
the background job.