Filter content items by departments
As a content designer I want to filter by departments so that I can see
the content that is most relevant to me.

Here, updates have been made to present a content items filter form with a
select box containing a list of the existing organisations, hence making
it possible to filter content items by an organisation.

This is presented on the "/content_items/filter" path. This is understood to
be the temporary and will change. This also reflects certain placements of
the tests within related content items files, these are also temporary and
will be moved once much progress is made.

Controller, feature, route and view specs have been added to further
describe the expected behaviour.

Bootstrap [1] form css classes / layout have been used to give some
initial presentation context to the filtering form.

[1] http://getbootstrap.com/css/#forms