Add support for Group models
A group is a set of content items with no predefined order. It has the
following properties:

- slug
- name
- group_type
- parent_group_id

So, groups have a tree structure, with no limit in the number of
descendants.

Groups also have a slug which is uniq.
The ContentIDs that belong to a Group has not beed added yet as I would
prefer to limit the scope of this commit.