Show title based on multiple filters
Previously we only showed the title of the content items listing page to
be that of the organisation if we were filtering by that criteria. But
now we can filter by Taxonomy as well and it is any combination of the
two. This should really be reflected to the user otherwise there is no
indication other than in the URL.