Add parent group to Endpoint to create Groups
Groups are hierarchical, so we should be able to assign a parent group
to any group. We have decided to use the Group slug to assign the
parent, because it seems easier to remember to the API user than the
Group ID, which is a number that is internal to the database.

```terminal
curl -X "POST" "http://content-performance-manager.dev.gov.uk/groups.json" \
     -H "Content-Type: application/json" \
     -d $'{
  "group": {
    "slug": "the-slug",
    "group_type": "the-group-type",
    "name": "the-name",
    "content_item_ids": [
      "content-id-1",
      "content-id-2"
    ],
    "parent_group_slug": "firstname49"
  },
  "api_token": "1234567"
}'
```