Change from id param to content_id
Instead of passing the internal database ID for a content item, change to
use the content_id as provided by the publishing api. This is in
preparation for subsequent changes that will implement an API endpoint
to provide content item details. Rather than accepting an ID param for
that feature, we will use the content_id.

This is also consistent with our future intention to have less of a full
representation of all content items in our app database and rely more on
the publishing platform APIs.