Fix missing taxonomy content id when sorting
Fixes bug where sorting the list of content items when a taxonomy is
applied removes the taxonomy filtering. This was because we weren’t
including the taxonomy content id in the URL for each of the sorting
option links.

This is a first pass to fix the bug. Subsequent work needed to refactor.