Replace view test with feature test
Use a feature test to validate that the content item renders the
organisations it belongs to instead of the view test.

Some view tests were repeated in the feature tests, which we have now
removed; this also indicated that the view tests should be removed
altogether as the feature tests give us enough confidence in our test
suite.

As a side effect, the feature tests are also more concise.