Move test for not published items into feature tests
As this is the last test of the file, the `show.html.erb_spec.rb` can be
safely removed.

Some view tests were repeated in the feature tests, which we have now
removed; this also indicated that the view tests should be removed
altogether as the feature tests give us enough confidence in our test
suite.

As a side effect, the feature tests are also more concise.