Remove unnecessary view tests
These tests are already covered in:

- features/filter_spec.rb
- features/summary_spec.rb

Some view tests were repeated in the feature tests, which we have now
removed; this also indicated that the view tests should be removed
altogether as the feature tests give us enough confidence in our test
suite.

As a side effect, the feature tests are also more concise.