Remove organisations list feature
Listing all the organisations along with the number of content item
for them is a feature that is no longer needed. This feature was
introduced as a way to exercise user stories that dealth with
organisations but after many user research sessions the view has proved
to be not very useful.