Use "topics" to be consistent with Whitehall
In the Whitehall user interface taxons are called "topics". To
differentiate them from the other topics (https://www.gov.uk/topic)
we've put in "new taxonomy" in parentheses in some places.

Since the users of Whitehall & CPM will be the same, we should probably
use the same terminology across applications.