Test that implicit building of responses works
When assigning `questions` to the `audit`, Rails
notices this is a `has_many :through` association
and automatically builds responses for us.

For this to work, we need to tell Rails the name
of the method to set the audit on responses, which
we do with the `inverse_of: :audit` option.