Tidy and simplify rails_helper
Remove lots of boilerplate so it’s easier to see
what’s going on. Basically, we need to clean the
database by truncating all the tables if the
webdriver runs in a separate process to the
database connection.

If we truncate the database we need to ensure we
recreate the seed data.