Define Capybara module so deployments do not break
When running deploy to staging via command[1], we were getting a 
deployment error because Capybara module couldn’t be found.

It seems that Capybara has not been loaded yet, hence the module can’t
be found. We can fix this issue by defining the module explicitly.

This is the build error that triggered this PR[2]


[1]: govuk_setenv content-performance-manager bundle exec rake RAILS_ENV=production RAILS_GROUPS=assets assets:precompile
[2]: https://deploy.staging.publishing.service.gov.uk/job/Deploy_App/3763/console