Include subpaths when querying GA for Unique Visits
Guidances has many pages, but only the root one is represented in the
publishing-api. Each page within the Guidance is a subpath from the
main path.

As a result, when query for stats, we are not getting aggregate values,
just the number of unique visitors to the index page (which is not
correct)