Continue to next item on successful save
This was a bit fiddly to get working properly.

1) We pass the filter_params through in the form
url so we reconstruct the scope correctly in the
#save action to determine the next item.

2) We check if there’s a next_item if the audit
was updated successfully and redirect to that.

Note: We need to make sure #next_item is always
called in #save so we set the instance variable
For the `else` branch. This is covered by the
Feature test, but it’s not immediately obvious
that we depend upon this side-effect.