Rename "Taxonomy" to "Taxon"
The word "taxonomy" describes the structure (as in "we're working
towards a single-subject taxonomy"). We do have multiple taxonomies on
GOV.UK at the moment, like the "specialist sector" taxonomy
(https://www.gov.uk/topic), the browse page taxonomy
(https://www.gov.uk/browse).

The items in the taxonomy we call "taxons" in code and content tagger,
and "topics" in some publisher applications. This commit renames
"Taxonomy" so that it's clear we're talking about the same thing.