Create TaxonomyProject and TaxonomyTodo models
We're currently working to create a new feature in this app that will
allow content designers to generate terms for a set of pages. The terms
will be used to create a new branch of the taxonomy (we're currently
working on environment and transport related content).

This commit creates two new models, which we will use in subsequent
pull requests (one to create an importer to create todos, the other one
with the term generation interface).