Create an inventory management page
We need to create rules for themes and sub themes
to be able to filter by them. We could have done
this by parsing the spreadsheet or in a rails
console, but this approach seems better because
this might actually be valuable to use in the
future.

Currently validation errors aren’t handled. For
example, two Themes with he same name will result
in a 500 server error.