Simplify RSpec factories
I don’t think we need separate files for
factories. We only have a ~100 lines in total and
half of that is going away soon when we drop
the organisations, taxons and groups.

Also, I’ve moved the test factory for google
analytics out of the GoogleAnalytics namespace. I
think this way makes it clearer this is a
test-only module.