Calculate number of PDFs based on `details` field
We are not storing properly the number of Pdfs for Content Items. After
a few refactors we removed the `details` attribute from the API so
we were not counting PDF

We are storing the content of a Content Item in the attribute `details`.
I am not storing it in the database because I can’t see now a real need
to do it, and I prefer to postpone this decision for a later moment.

Metrics are calculated based on the ContentItem object and no on a Hash
of attributes, which is in itself another improvement.