Precompute report rows
Report generation was taking too long. This change
precomputes row data into a table ahead of time
when saving the content item / links / audit. The
report can then just look up these rows which is
much faster.

The queries were taking a long time because they
had to join across multiple tables. Specifically,
we had to join from the filtered content items to
the audits and their responses. We also had to
join through the links table back to content items
to get the Primary Org. and Other Orgs. data.

After merging, the report_rows table will be
empty. We need to wait for the import of
Publishing API data to run before this will be
set. If we need to run this manually, we can use
the `rake report:precompute` rake task. We'll
need to do this if we add / delete questions to
refresh the precomputed rows.