Change Next to redirect to the next item
Rather than rending the audit page with the direct link to the next
item. This should provide a better experience for a individual user
performing audits, in the case when the states of other audits are
changing.

For example, when deciding the next item upon rendering the audit
page, if that next item is audited by another user in the mean time,
that link becomes stale. Deferring deciding the next item until the
last possible moment will help prevent this from happening.

A couple of tests expected that the Next link would not appear
sometimes. As finding out this information has been deferred, these
expectations have been removed.